move pages dir from src/pages to root
